Extract grouped incremental execution
